requirements.txt
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
transformers==4.44.0
datasets==2.21.0
librosa==0.10.1
opencv-python==4.10.0.84
scikit-learn==1.5.1
lightgbm==4.5.0
streamlit==1.38.0
fastapi==0.115.0
uvicorn==0.30.6
pillow==10.4.0
numpy==2.1.1
pandas==2.2.2
matplotlib==3.9.2
seaborn==0.13.2
tqdm==4.66.5
accelerate==1.0.1
bitsandbytes==0.43.3
soundfile==0.12.1
resampy==0.4.3
